fix initrd repair function
The method used to extract the service name in the fix_dump_comp_initrd
function is invalid because the get_service method is not part of the
check object provided to fix_dump_comp_initrd function.

There is no way to figure out the service name from the check object
provided to fix_dump_comp_initrd. Hence updating the function signature
to enforce the function caller to pass the service name.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>