Allow distro classes to redefine the capture kernel memory reservation
The amount of memory reservation needed for the capture kernel is dependent on
the distro type. But the variable that defines the capture kernel memory
reservation is defined inside a function that restricts the distro-specific
classes to redefine the capture kernel memory reservations.

This patch moves the variable used to define the capture kernel memory
reservation table to a class level so that distro classes can redefine it.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>