Add active dump check in kdump and FADump plugin
When the kernel panics it generates /proc/vmcore file that contains
physical memory of crashed kernel. If the kdump/FADump is configured,
system boots with secondary/capture kernel and detects the /proc/vmcore
file to collect the dump. Now for some reason if kdump service failed to
collect dump the it remains active in the system until we reboot the system.

If system has an active dump then it may create problem during next kernel
crash. This patch adds a check in kdump and fadump plugin to ensure that
there is no active dump in the system and if found raise flag to reboot
the system.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>