Take crashkernel recommendation from kdump-lib.sh scripts
RHEL has introduced a new function in kdump-lib.sh that provides crashkernel
recommendation for current system based on the system memory.

Now we take crashkernel value recommendation from the new function for
RHEL and Fedora. Also to maintain the backward compatibility if the new
crashkernel recommendation function is not present in kdump-lib.sh we will
use inbuilt crashkernel values.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>