[Daemon] check active status along with enabled
The disabled/enabled option of systemctl only controls whether a daemon
should start at boot or not. But to ensure the daemon runs all the time,
we need to check the active status of the daemon as well.

To achieve this the daemon check function is updated to validate the active
status of the daemon along with disabled/enabled. Similarly, the daemon
repair function is also updated to activate/start the daemon if reported
inactive by check function.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>