[spyre] reload udev rules to fix vfio permissions
There is already a check to ensure that the correct permissions are
configured in the vfio udev rules.

So, avoid setting vfio permissions manually with chmod or chown.
Instead, reload the udev rules to fix the vfio device permissions.

The udevadm <control | trigger | settle> commands are used to adjust
the vfio device permissions.

Signed-off-by: Sahithi Ravindranath <Sahithi.Ravindranath@ibm.com>
Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>