[spyre] add correct rule and remove redundant vfio rules
vfio_udev rule files fully has to be validated in case its mode and
permission is over written with later on rule and repair of udev rules
doesnt remove the redundant and old udev rules.

As part of fix adding persistant udev rule at begining of file
such that it not overwritten and remove redundant udev rules added by
olfer version of servicereport.

Signed-off-by: Sahithi Ravindranath <Sahithi.Ravindranath@ibm.com>
Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>